fix(db): 서버 SQL에서 PostgreSQL :: 캐스트 제거 및 CAST/암시적 바인딩으로 통일
- initDb 인덱스·DEFAULT·COUNT·regclass 등 전부 :: 없이 정리 (analytics_events 날짜 인덱스 포함)

- index, analytics, consent, insurer news, customer/sms/seed/스크립트 등 server 전역 동일 적용

Made-with: Cursor

+++ server/apis/customerExtraApi.js
@@ -171,7 +171,7 @@ export function registerCustomerExtraApi(apiRouter, ctx) {
       const r = await safeQuery(
         pool,
         `
-        SELECT customer_id, COUNT(*)::int AS c
+        SELECT customer_id, COUNT(*) AS c
         FROM customer_consultations
         WHERE user_id = $1 AND ga_id = $2
         GROUP BY customer_id
